bump A.B, i.e. major version according to PVP.
> If any entity was removed, or the types of any entities or the
> definitions of datatypes or classes were changed, or orphan instances
> were added or any instances were removed, then the new A.B must be
> greater than the previous A.B. Note that modifying imports or depending
> on a newer version of another package may cause extra orphan instances
> to be exported and thus force a major version change.
>
> https://wiki.haskell.org/Package_versioning_policy

i really hope i've got this right but i suspect i haven't. according to
the PVP entry on the haskell wiki, the first two parts of the version
constitute a major version. since this is a big refactoring of the
library, i believe this to be a major version bump. therefore i've
updated to bump from 0.0.1 -> 0.1.0.